Separate themes into own namespace and files
